requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
certifi==2020.4.5.2
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.7.0rc1
Django==2.2.7
django-braces==1.14.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-jet==1.0.8
django-modeltranslation==0.15
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
django-rest-multiple-models==2.1.3
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-yasg==1.17.1
google-api-python-client==1.4.1
gunicorn==20.0.4
hiredis==1.0.1
httplib2==0.18.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==1.0.0
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.4
Pillow==7.1.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
zope.interface==5.1.0
